+++ packages/mcu-debug/src/frontend/extension.ts
@@ -30,7 +30,7 @@ import { IOTerminal } from "./io-terminal";
 import { GDBServerConsole } from "./server-console";
 import { CDebugSession, CDebugChainedSessionItem } from "../common/cli-session";
 import { ServerConsoleLog } from "../adapter/server-console-log";
-import { logger } from '../common/cli-logger';
+import { logger } from '../common/logger';
 import { VscodeOutputChannelTransport } from './vscode-transport';
 import { isVarRefGlobalOrStatic } from "../adapter/var-scopes";
 import { getWSLNetworkingMode, ProvisioningResults, ProxyProvisionRequest, setDevelopmentModeEnvVars } from "@mcu-debug/shared";